+++ src/main/java/com/att/tdp/bisbis10/repository/DishRepository.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Repository;
 import com.att.tdp.bisbis10.entity.Dish;
 
 @Repository
-public interface DishRepository extends JpaRepository<Dish, Long> {
+public interface DishRepository extends JpaRepository<Dish, Integer> {
 
     @Query("SELECT d FROM Dish d WHERE d.restaurant.id = ?1")
-    public List<Dish> getDishesByRestaurantId(Long restaurantId);
+    public List<Dish> getDishesByRestaurantId(Integer restaurantId);
 }
