+++ src/main/java/com/att/tdp/bisbis10/repository/RestaurantRatingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.att.tdp.bisbis10.entity.RestaurantRating;
 
-public interface RestaurantRatingRepository extends JpaRepository<RestaurantRating, Long> {
+public interface RestaurantRatingRepository extends JpaRepository<RestaurantRating, Integer> {
 
 }
